refactoring: moved CommentCard from UI to Comments in HomePage, because it using only here
